# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Buyurtma Boshqaruv Tizimi API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface StoreLoginRequest {
  username: string;
  password: string;
}

export interface Store {
  id: number;
  name: string;
  username: string;
  createdAt: string;
}

export interface StoreLoginResponse {
  token: string;
  store?: Store;
  role: string;
}

export interface PinLoginRequest {
  pin: string;
  storeId: number;
}

export type AccountRole = (typeof AccountRole)[keyof typeof AccountRole];

export const AccountRole = {
  superadmin: "superadmin",
  admin: "admin",
  viewer: "viewer",
  worker: "worker",
} as const;

export interface Account {
  id: number;
  name: string;
  role: AccountRole;
  pin?: string | null;
  storeId?: number | null;
  storeName?: string | null;
  createdAt: string;
}

export interface PinLoginResponse {
  token: string;
  account: Account;
  role: string;
}

export interface MeResponse {
  id: number;
  username: string;
  role: string;
  storeId?: number | null;
  storeName?: string | null;
  name?: string | null;
}

export interface CreateStoreRequest {
  name: string;
  username: string;
  password: string;
}

export type CreateAccountRequestRole =
  (typeof CreateAccountRequestRole)[keyof typeof CreateAccountRequestRole];

export const CreateAccountRequestRole = {
  superadmin: "superadmin",
  admin: "admin",
  viewer: "viewer",
  worker: "worker",
} as const;

export interface CreateAccountRequest {
  name: string;
  role: CreateAccountRequestRole;
  pin?: string | null;
  storeId?: number | null;
}

export type UpdateAccountRequestRole =
  (typeof UpdateAccountRequestRole)[keyof typeof UpdateAccountRequestRole];

export const UpdateAccountRequestRole = {
  superadmin: "superadmin",
  admin: "admin",
  viewer: "viewer",
  worker: "worker",
} as const;

export interface UpdateAccountRequest {
  name?: string;
  role?: UpdateAccountRequestRole;
  pin?: string | null;
  storeId?: number | null;
}

export interface ServiceType {
  id: number;
  name: string;
  storeId?: number | null;
  createdAt: string;
}

export interface CreateServiceTypeRequest {
  name: string;
  storeId?: number | null;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  new: "new",
  accepted: "accepted",
  ready: "ready",
} as const;

export interface Order {
  id: number;
  orderId: string;
  status: OrderStatus;
  serviceTypeName: string;
  quantity: number;
  unit?: string | null;
  shelf?: string | null;
  notes?: string | null;
  storeId: number;
  storeName: string;
  clientId?: number | null;
  clientName?: string | null;
  clientPhone?: string | null;
  createdByName: string;
  acceptedByName?: string | null;
  acceptedAt?: string | null;
  readyAt?: string | null;
  createdAt: string;
}

export interface CreateOrderRequest {
  serviceTypeId: number;
  quantity: number;
  unit?: string | null;
  shelf?: string | null;
  notes?: string | null;
  clientId?: number | null;
  clientName?: string | null;
  clientPhone?: string | null;
}

export type UpdateOrderStatusRequestStatus =
  (typeof UpdateOrderStatusRequestStatus)[keyof typeof UpdateOrderStatusRequestStatus];

export const UpdateOrderStatusRequestStatus = {
  new: "new",
  accepted: "accepted",
  ready: "ready",
} as const;

export interface UpdateOrderStatusRequest {
  status: UpdateOrderStatusRequestStatus;
}

export interface OrdersSummary {
  new: number;
  accepted: number;
  ready: number;
  totalToday: number;
}

export type ClientStatus = (typeof ClientStatus)[keyof typeof ClientStatus];

export const ClientStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Client {
  id: number;
  firstName: string;
  lastName: string;
  phone: string;
  telegramUserId?: string | null;
  status: ClientStatus;
  createdAt: string;
}

export interface CreateClientRequest {
  firstName: string;
  lastName: string;
  phone: string;
  telegramUserId?: string | null;
}

export type GetOrdersParams = {
  status?: GetOrdersStatus;
  storeId?: number;
  date?: string;
  search?: string;
};

export type GetOrdersStatus =
  (typeof GetOrdersStatus)[keyof typeof GetOrdersStatus];

export const GetOrdersStatus = {
  new: "new",
  accepted: "accepted",
  ready: "ready",
} as const;

export type GetClientsParams = {
  search?: string;
  status?: GetClientsStatus;
};

export type GetClientsStatus =
  (typeof GetClientsStatus)[keyof typeof GetClientsStatus];

export const GetClientsStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;