# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Buyurtma Boshqaruv Tizimi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/store-login:
    post:
      operationId: storeLogin
      tags: [auth]
      summary: Login with store username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreLoginRequest"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreLoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/pin-login:
    post:
      operationId: pinLogin
      tags: [auth]
      summary: Login with PIN code (for workers)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinLoginRequest"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinLoginResponse"
        "401":
          description: Invalid PIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /accounts:
    get:
      operationId: getAccounts
      tags: [accounts]
      summary: Get all accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    put:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /clients:
    get:
      operationId: getClients
      tags: [clients]
      summary: Get all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Manually create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /clients/{id}/approve:
    post:
      operationId: approveClient
      tags: [clients]
      summary: Approve a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}/reject:
    post:
      operationId: rejectClient
      tags: [clients]
      summary: Reject a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    StoreLoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    StoreLoginResponse:
      type: object
      properties:
        token:
          type: string
        store:
          $ref: "#/components/schemas/Store"
        role:
          type: string
      required: [token, role]
    PinLoginRequest:
      type: object
      properties:
        pin:
          type: string
        storeId:
          type: integer
      required: [pin, storeId]
    PinLoginResponse:
      type: object
      properties:
        token:
          type: string
        account:
          $ref: "#/components/schemas/Account"
        role:
          type: string
      required: [token, account, role]
    MeResponse:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        storeId:
          type: integer
          nullable: true
        storeName:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
      required: [id, username, role]
    Store:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, username, createdAt]
...
    Account:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
          enum: [superadmin, admin, viewer, worker]
        pin:
          type: string
          nullable: true
        storeId:
          type: integer
          nullable: true
        storeName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, role, createdAt]
    CreateAccountRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [superadmin, admin, viewer, worker]
        pin:
          type: string
          nullable: true
        storeId:
          type: integer
          nullable: true
      required: [name, role]
    UpdateAccountRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [superadmin, admin, viewer, worker]
        pin:
          type: string
          nullable: true
        storeId:
          type: integer
          nullable: true
...
    Client:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        telegramUserId:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time
      required: [id, firstName, lastName, phone, status, createdAt]
    CreateClientRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        telegramUserId:
          type: string
          nullable: true
      required: [firstName, lastName, phone]