# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Buyurtma Boshqaruv Tizimi API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountRole } from "./accountRole";

export interface Account {
  id: number;
  name: string;
  role: AccountRole;
  pin?: string | null;
  storeId?: number | null;
  storeName?: string | null;
  createdAt: Date;
}