# lib/api-zod/src/generated/types/pinLoginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Buyurtma Boshqaruv Tizimi API
 * OpenAPI spec version: 0.1.0
 */

export interface PinLoginRequest {
  pin: string;
  storeId: number;
}