# lib/db/src/schema/accounts.ts
import { pgTable, serial, text, integer, timestamp, pgEnum } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { storesTable } from "./stores";

export const accountRoleEnum = pgEnum("account_role", ["sudo", "superadmin", "admin", "viewer", "worker"]);

export const accountsTable = pgTable("accounts", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  role: accountRoleEnum("role").notNull(),
  pin: text("pin"),
  username: text("username").unique(),
  passwordHash: text("password_hash"),
  storeId: integer("store_id").references(() => storesTable.id, { onDelete: "cascade" }),
  createdAt: timestamp("created_at").notNull().defaultNow(),
});

export const insertAccountSchema = createInsertSchema(accountsTable).omit({ id: true, createdAt: true });
export type InsertAccount = z.infer<typeof insertAccountSchema>;
export type Account = typeof accountsTable.$inferSelect;